+++ src/components/Meals/AvailableMeals.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Card from "../UI/Card/Card";
+import Card from "../UI/Card";
 
 import styles from "./AvailableMeals.module.css";
 import MealItem from "./MealItem/MealItem";
