+++ src/components/UI/Input.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+import styles from "./Input.module.css";
+
+const Input = (props) => {
+  return (
+    <div className={styles.input}>
+      <label htmlFor={props.input.id}>{props.label}</label>
+      <input {...props.input} />
+    </div>
+  );
+};
+
+export default Input;
